requirements.txt
fastapi==0.109.2
pip-chill==1.0.3
python-decouple==3.8
python-multipart==0.0.9
uvicorn==0.27.1
